railway-nba-api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
nba_api>=1.4.1
psycopg2-binary>=2.9.9
pandas>=2.0.0
